+++ src/index.ts
@@ -6,5 +6,5 @@ export {Collection} from "./util/Collection";
 //export {Setup as BaseServiceWorkerSetup} from "./services/BaseServiceWorker";
 //export {Setup as IPCSetup} from "./util/IPC";
 export {Options, Stats, ClusterStats, ShardStats, ServiceStats, ObjectLog, StartingStatus, ReshardOptions, LoggingOptions, ClusterCollection, ServiceCollection, ServiceCreator} from "./sharding/Admiral";
-export {IPC} from "./util/IPC";
+export {IPC, CentralStore} from "./util/IPC";
 //export {Queue, QueueItem, ClusterConnectMessage, ServiceConnectMessage, ShutdownMessage} from "./util/Queue";